Move commitChange method to an activity extension
This allows us to re-use the code for other classes.
